+++ dmarc-report/dmarc-report.d/local/share/cook/bin/configure-parsedmarc.sh
@@ -16,6 +16,10 @@ if ! id -u "parsedmarc" >/dev/null 2>&1; then
   /usr/sbin/pw useradd -n parsedmarc -c 'parsedmarc user' -m -d /opt/parsedmarc -s /bin/sh -h -
 fi
 
+# create a folder for output reports in json and csv
+mkdir -p "/mnt/$OUTPUTFOLDER"
+chown -R parsedmarc "/mnt/$OUTPUTFOLDER"
+
 # create virtualenv
 sudo -u parsedmarc virtualenv /opt/parsedmarc/venv || true
 
@@ -36,12 +40,21 @@ TEMPLATEPATH=$(dirname "$SCRIPT")/../templates
 # safe(r) separator for sed
 sep=$'\001'
 
-## copy in custom .env file
-#< "$TEMPLATEPATH/env.in" \
-#  sed "s${sep}%%imapserver%%${sep}$IMAPSERVER${sep}g" | \
-#  sed "s${sep}%%imapuser%%${sep}$IMAPUSER${sep}g" | \
-#  sed "s${sep}%%imappass%%${sep}$IMAPPASS${sep}g" | \
-#  sed "s${sep}%%imapfolder%%${sep}$IMAPFOLDER${sep}g" | \
-#  sed "s${sep}%%serverport%%${sep}$SERVERPORT${sep}g" \
-#  > /root/dmarc-report/.env
+# copy in custom parsedmarc.ini file
+< "$TEMPLATEPATH/parsedmarc.ini.in" \
+  sed "s${sep}%%imapserver%%${sep}$IMAPSERVER${sep}g" | \
+  sed "s${sep}%%imapuser%%${sep}$IMAPUSER${sep}g" | \
+  sed "s${sep}%%imappass%%${sep}$IMAPPASS${sep}g" | \
+  sed "s${sep}%%imapfolder%%${sep}$IMAPFOLDER${sep}g" | \
+  sed "s${sep}%%outputfolder%%${sep}$OUTPUTFOLDER${sep}g" \
+  > /usr/local/etc/parsedmarc.ini
+
+# need steps to run python process to get report
+#
+# as user parsedmarc run
+# /opt/parsedmarc/venv/bin/parsedmarc -c /usr/local/etc/parsedmarc.ini
+#
+# this takes a long time to complete so should not be a pot startup command
+# rather a service that starts some other way and backgrounds after the pot
+# image is live